+++ examples/lib/widgetUI/demo_util/demo3_body.dart
@@ -23,7 +23,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_sound/flutter_sound.dart';
 import 'package:intl/date_symbol_data_local.dart';
-import 'package:flutter_sound_demo/demo_util/temp_file.dart';
+import 'package:flutter_sound_examples/widgetUI/demo_util/temp_file.dart';
+
 import 'package:permission_handler/permission_handler.dart';
 
 import 'demo_active_codec.dart';