+++ examples/lib/widgetUI/demo_util/recorder_state.dart
@@ -23,7 +23,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_sound/flutter_sound.dart';
 
 import 'package:flutter_sound/src/util/log.dart';
-import 'package:flutter_sound_demo/demo_util/temp_file.dart';
+import 'package:flutter_sound_examples/widgetUI/demo_util/temp_file.dart';
 import 'demo_active_codec.dart';
 import 'demo_media_path.dart';